+++ R/add.R
@@ -31,6 +31,8 @@ add_insulin_purchases_cols <- function(gld_hba1c_after_drop_steps) {
     logic_as_expression()
 
   insulin_cols <- gld_hba1c_after_drop_steps |>
+    # Remove hba1c rows so dates from HbA1c measurements aren't included.
+    dplyr::filter(is.na(.data$is_hba1c)) |>
     # `volume` is the doses contained in the purchased package and `apk` is the
     # number of packages purchased
     dplyr::mutate(
@@ -145,8 +147,7 @@ add_t1d_diagnoses_cols <- function(data) {
         \(x) dplyr::coalesce(x, 0)
       )
     ) |>
-    # Keep earliest two dates per individual.
-    dplyr::filter(dplyr::row_number(.data$date) %in% 1:2, .by = "pnr") |>
+    # Add two columns for type 1 diabetes diagnoses logic.
     dplyr::mutate(
       has_majority_t1d_diagnoses = !!logic$has_majority_t1d_diagnoses,
       has_any_t1d_primary_diagnosis = !!logic$has_any_t1d_primary_diagnosis