+++ tests/testthat/test-drop.R
@@ -7,7 +7,6 @@ bef <- simulate_registers("bef", 1000)[[1]] |>
   )
 gld_purchases <- simulate_registers("lmdb", 1000)[[1]] |>
   keep_gld_purchases() |>
-  add_insulin_purchases_cols() |>
   dplyr::add_row(
     date = "20200101",
     atc = "A10BA02",
@@ -52,7 +51,6 @@ lpr3 <- prepare_lpr3(
 
 no_pcos <- register_data$lmdb |>
   keep_gld_purchases() |>
-  add_insulin_purchases_cols() |>
   drop_pcos(register_data$bef)
 
 preg_dates <- keep_pregnancy_dates(lpr2, lpr3)